(PC-31535) test(app): use of deleted global mocks in tests
