+++ scripts/noUncheckedIndexedAccess_snapshot.txt
@@ -60,16 +60,16 @@
 ./src/features/navigation/RootNavigator/linking/getNestedNavigationFromState.ts:14
 ./src/features/navigation/RootNavigator/linking/getNestedNavigationFromState.ts:16
 ./src/features/navigation/RootNavigator/linking/getNestedNavigationFromState.ts:19
-./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:74
-./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:80
-./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:86
-./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:92
-./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:98
+./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:75
+./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:81
+./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:87
+./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:93
+./src/features/navigation/RootNavigator/linking/getScreenConfig.test.ts:99
 ./src/features/navigation/helpers/useCurrentRoute.ts:4
 ./src/features/navigation/helpers/usePreviousRoute.ts:4
 ./src/features/navigation/sanitizeNavigationState.ts:5
-./src/features/offer/components/OfferPlaylistList/OfferPlaylistList.native.test.tsx:105
-./src/features/offer/components/OfferPlaylistList/OfferPlaylistList.native.test.tsx:71
+./src/features/offer/components/OfferPlaylistList/OfferPlaylistList.native.test.tsx:121
+./src/features/offer/components/OfferPlaylistList/OfferPlaylistList.native.test.tsx:87
 ./src/features/offer/components/VenueSelectionListItem/VenueSelectionListItem.web.tsx:74
 ./src/features/offer/components/VenueSelectionListItem/VenueSelectionListItem.web.tsx:76
 ./src/features/offer/components/VenueSelectionListItem/VenueSelectionListItem.web.tsx:81
