+++ src/features/bookOffer/pages/BookingOfferModal.web.test.tsx
@@ -59,6 +59,7 @@ jest.mock('api/useSearchVenuesOffer/useSearchVenueOffers', () => ({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('<BookingOfferModal/>', () => {
   describe('Accessibility', () => {
