+++ src/features/gtlPlaylist/components/GtlPlaylist.native.test.tsx
@@ -44,6 +44,22 @@ jest.mock('react-native-intersection-observer', () => {
 
 jest.spyOn(useFeatureFlagAPI, 'useFeatureFlag').mockReturnValue(false)
 
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
+})
+
 describe('GtlPlaylist', () => {
   describe('on venue page', () => {
     it('should log ConsultOffer when pressing an item', () => {
