+++ src/features/home/components/modules/HomeModule.native.test.tsx
@@ -103,6 +103,21 @@ const defaultDataVenues: ModuleData = {
 }
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
+})
 
 describe('<HomeModule />', () => {
   beforeEach(() => {
