+++ src/features/home/components/modules/OffersModule.native.test.tsx
@@ -47,6 +47,21 @@ jest.mock('features/auth/context/AuthContext')
 jest.mock('libs/subcategories/useSubcategories')
 
 jest.spyOn(useFeatureFlagAPI, 'useFeatureFlag').mockReturnValue(false)
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
+})
 
 describe('OffersModule', () => {
   it('should not render if data is undefined', () => {
