+++ src/features/identityCheck/pages/identification/educonnect/EduConnectForm.web.test.tsx
@@ -13,6 +13,7 @@ jest.spyOn(useEduConnectLoginAPI, 'useEduConnectLogin').mockReturnValue({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('<EduConnectForm />', () => {
   it('should render EduConnectForm', () => {
