+++ src/features/identityCheck/pages/profile/SetAddress.web.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { SetAddress } from 'features/identityCheck/pages/profile/SetAddress'
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
-import { waitFor, screen, checkAccessibilityFor, render } from 'tests/utils/web'
+import { checkAccessibilityFor, render, screen, waitFor } from 'tests/utils/web'
 
 jest.mock('uuid', () => ({
   v1: jest.fn(),
@@ -10,6 +10,7 @@ jest.mock('uuid', () => ({
 }))
 
 jest.mock('features/identityCheck/context/SubscriptionContextProvider')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('<SetAddress/>', () => {
   it('should render correctly', () => {
