+++ src/features/internal/marketingAndCommunication/components/DeeplinksGeneratorForm.native.test.tsx
@@ -14,6 +14,7 @@ jest.mock('libs/subcategories/useSubcategories')
 mockdate.set(new Date('2022-08-09T00:00:00Z'))
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<DeeplinksGeneratorForm />', () => {
   it('should render deeplink generator form with marketing as default utm_gen', () => {
