+++ src/features/offer/components/OfferPlaylist/OfferPlaylist.native.test.tsx
@@ -9,6 +9,22 @@ const renderItemMock = jest.fn().mockReturnValue(null)
 const itemWidth = 100
 const itemHeight = 50
 
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
+})
+
 describe('<OfferPlaylist />', () => {
   it('should display correctly the playlist', async () => {
     render(
