+++ src/features/offer/components/OfferPlaylistList/OfferPlaylistList.native.test.tsx
@@ -44,6 +44,22 @@ const defaultRemoteConfig: CustomRemoteConfig = {
   sameAuthorPlaylist: 'sameAuthorPlaylist',
 }
 
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
+})
+
 describe('<OfferPlaylistList />', () => {
   describe('Similar offers', () => {
     describe('Same category playlist', () => {
