+++ src/features/profile/pages/ChangeCity/ChangeCity.web.test.tsx
@@ -11,6 +11,7 @@ jest.mock('uuid', () => ({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('<ChangeCity/>', () => {
   it('should render correctly', () => {
