+++ src/features/profile/pages/NewEmailSelection/NewEmailSelection.web.test.tsx
@@ -12,6 +12,7 @@ jest.mock('uuid', () => {
     v4: jest.fn(() => value++),
   }
 })
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<NewEmailSelection />', () => {
   describe('Accessibility', () => {
