+++ src/features/profile/pages/NotificationSettings/NotificationsSettings.web.test.tsx
@@ -24,6 +24,7 @@ jest.spyOn(usePushPermission, 'usePushPermission').mockReturnValue({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('NotificationsSettings', () => {
   it('should render correctly', () => {
