+++ src/features/profile/pages/Profile.web.test.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
-import { render, checkAccessibilityFor, screen } from 'tests/utils/web'
+import { checkAccessibilityFor, render, screen } from 'tests/utils/web'
 import * as useVersion from 'ui/hooks/useVersion.web'
 
 import { Profile } from './Profile'
@@ -11,6 +11,7 @@ jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
 jest.mock('features/favorites/context/FavoritesWrapper')
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 jest.spyOn(useVersion, 'useVersion').mockReturnValue('Version\u00A01.10.5')
