+++ src/features/search/components/AutocompleteOffer/AutocompleteOffer.native.test.tsx
@@ -29,6 +29,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('AutocompleteOffer component', () => {
   describe('With suggestion hits', () => {
