+++ src/features/search/components/AutocompleteOfferItem/AutocompleteOfferItem.native.test.tsx
@@ -43,6 +43,7 @@ const mockSendEvent = jest.fn()
 const searchId = uuidv4()
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('AutocompleteOfferItem component', () => {
   beforeEach(() => {
