+++ src/features/search/components/CategoriesButtons/CategoriesButtons.native.test.tsx
@@ -21,6 +21,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('CategoriesButtons', () => {
   it('should display categories', async () => {
