+++ src/features/search/components/SearchHeader/SearchHeader.web.test.tsx
@@ -38,6 +38,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('SearchHeader component', () => {
   afterEach(() => {
