+++ src/features/search/components/SearchListHeader/SearchListHeader.web.test.tsx
@@ -42,6 +42,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
   useSearch: () => mockUseSearch(),
 }))
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('<SearchListHeader />', () => {
   describe('When wipVenueMap feature flag activated', () => {
