+++ src/features/search/components/SearchResultsContent/SearchResultsContent.native.test.tsx
@@ -152,6 +152,22 @@ jest.useFakeTimers()
 
 jest.mock('libs/firebase/analytics/analytics')
 
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
+})
+
 describe('SearchResultsContent component', () => {
   beforeAll(() => {
     mockHits = []
