+++ src/features/search/pages/Search/SearchN1/SearchN1.native.test.tsx
@@ -17,6 +17,7 @@ const mockSearchState = initialSearchState
 const mockDispatch = jest.fn()
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 jest.spyOn(useFeatureFlagAPI, 'useFeatureFlag').mockReturnValue(false)
 
 jest
@@ -30,6 +31,22 @@ jest.spyOn(useSearch, 'useSearch').mockReturnValue({
   isFocusOnSuggestions: false,
   showSuggestions: jest.fn(),
   hideSuggestions: jest.fn(),
+})
+
+jest.mock('@shopify/flash-list', () => {
+  const ActualFlashList = jest.requireActual('@shopify/flash-list').FlashList
+  class MockFlashList extends ActualFlashList {
+    componentDidMount() {
+      super.componentDidMount()
+      this.rlvRef?._scrollComponent?._scrollViewRef?.props?.onLayout({
+        nativeEvent: { layout: { height: 250, width: 800 } },
+      })
+    }
+  }
+  return {
+    ...jest.requireActual('@shopify/flash-list'),
+    FlashList: MockFlashList,
+  }
 })
 
 describe('<SearchN1/>', () => {
