+++ src/features/search/pages/SearchLanding/SearchLanding.native.test.tsx
@@ -218,6 +218,7 @@ const mockedEmptyHistory = {
 }
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<SearchLanding />', () => {
   beforeEach(() => {
