+++ src/features/search/pages/SearchLanding/SearchLanding.perf.test.tsx
@@ -12,6 +12,7 @@ import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { act, measurePerformance } from 'tests/utils'
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 jest.spyOn(useFeatureFlag, 'useFeatureFlag').mockReturnValue(true)
