+++ src/features/search/pages/SearchResults/SearchResults.native.test.tsx
@@ -211,6 +211,7 @@ const mockedEmptyHistory = {
 }
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<SearchResults/>', () => {
   mockUseNetInfoContext.mockReturnValue({ isConnected: true })
