+++ src/features/search/pages/SearchResults/SearchResults.perf.test.tsx
@@ -14,6 +14,7 @@ import { reactQueryProviderHOC } from 'tests/reactQueryProviderHOC'
 import { act, measurePerformance } from 'tests/utils'
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 jest.spyOn(useFeatureFlag, 'useFeatureFlag').mockReturnValue(false)
