+++ src/features/search/pages/SearchResults/SearchResults.web.test.tsx
@@ -84,6 +84,7 @@ const mockUseCenterOnLocation = useCenterOnLocation as jest.Mock
 
 jest.mock('features/venue/api/useVenueOffers')
 jest.mock('features/venueMap/helpers/zoomOutIfMapEmpty')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<SearchResults/>', () => {
   describe('Accessibility', () => {
