+++ src/features/subscription/page/OnboardingSubscription.web.test.tsx
@@ -6,6 +6,7 @@ import { act, checkAccessibilityFor, render, screen } from 'tests/utils/web'
 
 jest.mock('libs/firebase/analytics/analytics')
 jest.mock('libs/firebase/remoteConfig/remoteConfig.services')
+jest.mock('ui/theme/customFocusOutline/customFocusOutline')
 
 describe('OnboardingSubscription', () => {
   describe('Accessibility', () => {
