+++ src/ui/components/buttons/SubcategoryButton/SubcategoryButton.native.test.tsx
@@ -16,6 +16,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<SubcategoryButton/>', () => {
   it('should render SubcategoryButton', async () => {
