+++ src/ui/components/buttons/SubcategoryButton/SubcategoryButtonList.native.test.tsx
@@ -35,6 +35,7 @@ jest.mock('features/search/context/SearchWrapper', () => ({
 }))
 
 jest.mock('libs/firebase/analytics/analytics')
+jest.mock('features/navigation/TabBar/routes')
 
 describe('<SubcategoryButtonList/>', () => {
   it('should render SubcategoryButtonList', async () => {
